requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.20
django-environ==0.12.0
djangorestframework==3.16.0
idna==3.10
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
